refactor OTPHelpers: split utilities (implicits) and functions for otp generation
