+++ src/main/scala/it/unibo/authsim/library/dsl/otp/builders/HOTPBuilder.scala
@@ -2,8 +2,9 @@ package it.unibo.authsim.library.dsl.otp.builders
 
 import it.unibo.authsim.library.dsl.HashFunction
 import it.unibo.authsim.library.dsl.otp.builders.OTPBuilder.AbstractHOTPBuilder
+import it.unibo.authsim.library.dsl.otp.generator.OTPGenerator
 import it.unibo.authsim.library.dsl.otp.model.HOTP
-import it.unibo.authsim.library.dsl.otp.util.OTPHelpers.{generatorSeed, hmac, truncate}
+import it.unibo.authsim.library.dsl.otp.util.OTPHelpers.generatorSeed
 import it.unibo.authsim.library.dsl.policy.model.StringPolicies.OTPPolicy
 
 case class HOTPBuilder() extends AbstractHOTPBuilder:
@@ -14,8 +15,8 @@ case class HOTPBuilder() extends AbstractHOTPBuilder:
     override def hashFunction: HashFunction = HOTPBuilder.this._hashFunction
 
     override def policy: OTPPolicy = HOTPBuilder.this._policy
-    
-    override def generate: String = truncate(this.hashFunction, HOTPBuilder.this._secret, this.length, HOTPBuilder.this._seed)(hmac)
+
+    override def generate: String = OTPGenerator(this.hashFunction, HOTPBuilder.this._secret, this.length, HOTPBuilder.this._seed)
 
     override def check(pincode: String): Boolean =  this.generate == pincode
