+++ src/main/scala/it/unibo/authsim/library/dsl/otp/builders/TOTPBuilder.scala
@@ -2,8 +2,9 @@ package it.unibo.authsim.library.dsl.otp.builders
 
 import it.unibo.authsim.library.dsl.HashFunction
 import it.unibo.authsim.library.dsl.otp.builders.OTPBuilder.AbstractTOTPBuilder
+import it.unibo.authsim.library.dsl.otp.generator.OTPGenerator
 import it.unibo.authsim.library.dsl.otp.model.{HOTP, TOTP}
-import it.unibo.authsim.library.dsl.otp.util.OTPHelpers.{generatorSeed, hmac, truncate}
+import it.unibo.authsim.library.dsl.otp.util.OTPHelpers.generatorSeed
 import it.unibo.authsim.library.dsl.policy.model.StringPolicies.OTPPolicy
 
 import scala.concurrent.duration.Duration
@@ -25,14 +26,14 @@ case class TOTPBuilder() extends AbstractTOTPBuilder:
     override def policy: OTPPolicy = TOTPBuilder.this._policy
 
     override def generate: String =
-      val generated = truncate(this.hashFunction, TOTPBuilder.this._secret, this.length, TOTPBuilder.this._seed)(hmac)
+      val generated = OTPGenerator(this.hashFunction, TOTPBuilder.this._secret, this.length, TOTPBuilder.this._seed)
       this._createDate = Some(System.currentTimeMillis())
       generated
 
     override def check(pincode: String): Boolean =
       val validTime = this.createDate.isDefined && this.createDate.get + this.timeout.toMillis > System.currentTimeMillis()
       if !validTime then this.reset
-      validTime && truncate(this.hashFunction, TOTPBuilder.this._secret, pincode.length, TOTPBuilder.this._seed)(hmac) == pincode
+      validTime && OTPGenerator(this.hashFunction, TOTPBuilder.this._secret, pincode.length, TOTPBuilder.this._seed) == pincode
 
     override def reset: Unit = TOTPBuilder.this.generateSeed
