+++ src/main/scala/it/unibo/authsim/library/dsl/otp/util/OTPHelpers.scala
@@ -6,7 +6,6 @@ import it.unibo.authsim.library.dsl.policy.model.StringPolicies.OTPPolicy
 import javax.crypto.Mac
 import javax.crypto.spec.SecretKeySpec
 import scala.collection.immutable.NumericRange
-import scala.collection.mutable.Map as MutableMap
 import scala.language.postfixOps
 import scala.util.Random
 
@@ -41,38 +40,3 @@ object OTPHelpers:
   implicit val generatorSeed: () => Int = () => Random.between(Int.MinValue, Int.MaxValue)
 
   implicit val generatorLength: OTPPolicy => Int = (policy: OTPPolicy) => Random.between(policy.minimumLength, policy.maximumLength + 1)
-
-  def hmac(hashFunction: HashFunction, secret: String): Array[Byte] =
-    val algorithm: String = s"Hmac${hashFunction.getClass.getSimpleName}"
-    val mac: Mac = Mac.getInstance(algorithm)
-    mac.init(new SecretKeySpec(secret.getBytes, algorithm))
-    mac.doFinal(secret.getBytes)
-
-  private case class PreviousGenerateOTP(seed: Int = 0, start: Int = 0, pin: String = ""):
-    def isChangeSeed(value: Int): Boolean = this.seed != value
-    def regeneratedSamePin(pin: String): Boolean = this.pin == pin
-
-  private var previous: MutableMap[String, PreviousGenerateOTP] = MutableMap.empty // secret -> PreviousGenerateOTP
-
-  def truncate(hashFunction: HashFunction, secret: String, digits: Int, seed: Int)(hmacFunction: (HashFunction, String) => Array[Byte]): String =
-    val previousValues: PreviousGenerateOTP = previous.get(secret).getOrElse(PreviousGenerateOTP())
-//    println(s"\n\nPrevious = (seed = ${previousValues.seed}, start = ${previousValues.start}, pin = ${previousValues.pin})")
-    if previousValues.isChangeSeed(seed) then
-//      println(s"length of pin to generate = $digits")
-      val hmacStr: String = hmacFunction(hashFunction, secret).map(_.toUInt).mkString
-//      println(s"HMAC = $hmacStr, Len = ${hmacStr.length}")
-      val start: Int = Random.between(0, hmacStr.length, previousValues.start)
-//      println(s"Start: $start")
-      var pin = hmacStr.slice(start, start + digits)
-      if pin.length < digits then
-//        println(s"Add more char : '0'")
-        (1 to digits - pin.length).foreach(_ => pin = pin.appended('0'))
-      if previousValues.regeneratedSamePin(pin) then
-//        println(s"Change one char.")
-        pin = pin.change(('0' to '9'))
-      previous.update(secret, PreviousGenerateOTP(seed, start, pin))
-//      println(s"Now = (seed = $seed, start = $start, pin = $pin)\n\n")
-      pin
-    else
-//      println(s"get previous pin = ${previousValues.pin}")
-      previousValues.pin
